+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
+## [Unreleased]
+### Changed
+- `JSONExpect`: `count` may now be specified as an `IntRange`
+- `JSONExpect`: `length` checks the length of a string
+- `JSONExpect`: `uuid` checks a string as a UUID
+
 ## [0.3] - 2020-04-16
 ### Changed
 - `JSONExpect`: improved error messages